+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BoM:
   - Support for ${field} expansion. (#471)
   - LCSC links (SchrodingersGat/KiBoM#190)
+  - `parse_value` can be used to disable the *Value* parser (See #494)
+    Also added a warning about using extra data in the *Value* field.
 - iBoM:
   - `forced_name` option to force the name displayed at the top left corner
     (#470)